# Cache common/common_linux.Virtualization()