By assuming virtualization environment won't change during a the program's runtime, we can cache common/common_linux.Virtualization() with a simple map to reduce amount of system calls. I first mentioned this issue at [890](https://github.com/shirou/gopsutil/pull/890#issuecomment-690211919)