@AtakanColak 2020-09-11
I wanted to compare profiling results of this optimization. 

Before optimization of Virtualization(), there is a distinct compute weight on fillFromStatWithContext which calls Virtualization()

![image](https://user-images.githubusercontent.com/22523757/92925042-94d18780-f442-11ea-9e27-67e8a51deca4.png)

After optimization, we can clearly see that the compute weight is reduced and spread to the other functions. 

![image](https://user-images.githubusercontent.com/22523757/92924965-6eabe780-f442-11ea-9766-879a475e4cc5.png)