@shirou 2020-09-11
I think these kind of optimization should be implemented in Application side, not library side.