@AtakanColak 2020-09-11
@shirou could you elaborate on that? Most of the compute is spent with ReadLines on Virtualization which is in internal/common package. How could we optimize it from application side?