@shirou 2020-09-11
```
package main

import (
        "fmt"
        "github.com/shirou/gopsutil/host"
)

var (
        h = ""
        g = ""
)

func main() {
        if h == "" {
                h, g, _ := host.Virtualization()
                fmt.Println(h, g)
        } else {
                fmt.Println(h, g)
        }
}
```
or whatever, in your application.