@AtakanColak 2020-09-11
@shirou thank you for your reply. However your answer would be the case if it was my app who was calling Virtualization information, but it isn't. The function that calls common/common_linux.Virtualization() is fillFromStatWithContext() which is called by many process.Process methods in process_linux. We can't change private functions from the application side.