@Lomanic 2020-09-11
In fact the call to `common.Virtualization()` (should be VirtualizationWithContext() probably also) comes from [common.BootTimeWithContext()](https://github.com/shirou/gopsutil/blob/8a625ec054ea912247e0372de6b2a02255a6c09b/internal/common/common_linux.go#L58) which is not cached anymore since #857. As I wrote in https://github.com/shirou/gopsutil/pull/857#issuecomment-629858607, moving to unix.Sysinfo would maybe remove this call.

I fear caching this will make it harder to run unit tests for this function (we would have to hook to this new variable `virtualizationCache` I think).