@shirou 2020-09-12
ah, sorry for my misunderstanding. But as @Lomanic said, I also cache is not good for library. I will try to open a PR about using  `unix.Sysinfo`