@AtakanColak 2020-09-12
@Lomanic #857 mention a device can start with wrong date but the question we must ask for this case is can a device start with wrong virtualization data? I don't think it can but of course we must be certain.

I'm not going to comment on unit tests as I'm not sure on potential effects of caching on unit tests.