@Lomanic 2020-09-28
I think maybe we should cache this value in fact, for unit tests we can work around this by resetting this internal variable to nil if we want to call Virtualization() more than once. This would really improve performances for process.Processes() (see https://github.com/shirou/gopsutil/issues/842#issuecomment-697026725)