@Lomanic 2020-09-28
Here's a proper benchmarck

```bash
git -C "$(go env GOPATH)/src/github.com/shirou/gopsutil" checkout master
git -C "$(go env GOPATH)/src/github.com/shirou/gopsutil" reset --hard origin/master
git -C "$(go env GOPATH)/src/github.com/shirou/gopsutil" clean -f -d
cat <<EOF >> "$(go env GOPATH)/src/github.com/shirou/gopsutil/process/process_test.go"
func BenchmarkNewProcess(b *testing.B) {
	checkPid := os.Getpid()
	for i := 0; i < b.N; i++ {
		NewProcess(int32(checkPid))
    }
}
EOF
printf "\nBefore change:\n"
go test -bench=. github.com/shirou/gopsutil/process
curl -sL https://github.com/shirou/gopsutil/pull/942.diff | git apply
printf "\nAfter change:\n"
go test -bench=. github.com/shirou/gopsutil/process
```

Results are telling:
```
Before change:
goos: linux
goarch: amd64
pkg: github.com/shirou/gopsutil/process
BenchmarkNewProcess-4               2000            542412 ns/op
PASS
ok      github.com/shirou/gopsutil/process      3.106s

After change:
goos: linux
goarch: amd64
pkg: github.com/shirou/gopsutil/process
BenchmarkNewProcess-4              20000             78478 ns/op
PASS
ok      github.com/shirou/gopsutil/process      4.158s
```