@AtakanColak 2020-09-29
@Lomanic removed explicit nil as requested. Thank you for benchmarking results. Looking forward for the merge.