@reimda 2020-09-30
I tested this fix in telegraf and found that it makes performance similar to before #837 was merged.  Will this PR make it into a release soon?  We'd like to include it in telegraf 1.16.0 which is planned for next week.