@shirou 2020-10-01
@Lomanic sorry for interrupting your comment, but I think your request is resolved. So I merge this PR. Thank you for your review and contribution!