@shirou 2020-10-01
gopsutil is monthly release. After this PR merged, I have released `vv2.20.9`.